move private values to an env file
