+++ .env.example
@@ -1,0 +1,3 @@
+RPC_URL='YOUR_RPC_URL'
+PRIVATE_KEY='YOUR_PRIVATE_KEY'
+ACCOUNT_TYPE='simple' # simple, safe, kernel
