+++ .gitignore
@@ -1,2 +1,4 @@
 examples/*/node_modules/
 examples/*/yarn-error.log
+
+.env
