--- config.json
@@ -1,8 +1,0 @@
-{
-    "rpc_url": "YOUR_RPC_URL",
-    "private_key": "YOUR_PRIVATE_KEY",
-    "account_type": "simple",
-    "contract_address": "0x66519FCAee1Ed65bc9e0aCc25cCD900668D3eD49",
-    "function_name": "mintTo",
-    "entry_point": "0x5FF137D4b0FDCD49DcA30c7CF57E578a026d2789"
-}
