+++ package.json
@@ -1,0 +1,10 @@
+{
+  "name": "paymaster-bundler-examples",
+  "version": "1.0.0",
+  "main": "index.js",
+  "license": "MIT",
+  "dependencies": {
+    "dotenv": "^16.4.5"
+  },
+  "type": "module"
+}
